+++ packages/generator-one-app-module/generators/app/templates/base-child-module/src/index.js
@@ -1,3 +1,0 @@
-import ModuleContainer from './components/<%=moduleNamePascal%>';
-
-export default ModuleContainer;