+++ packages/one-app-bundler/webpack/module/webpack.client.js
@@ -24,7 +24,7 @@ const SriPlugin = require('webpack-subresource-integrity');
 const extendWebpackConfig = require('../../utils/extendWebpackConfig');
 const commonConfig = require('../webpack.common');
 const getConfigOptions = require('../../utils/getConfigOptions');
-
+require('../../utils/patchedCryptoHash');
 const {
   babelLoader,
   cssLoader,
@@ -55,7 +55,7 @@ module.exports = (babelEnv) => {
       resolve: {
         mainFields: ['browser', 'module', 'main'],
         modules: [packageRoot, 'node_modules'],
-        extensions: ['.js', '.jsx'],
+        extensions: ['.js', '.jsx', '.ts', '.tsx'],
       },
       node: { module: 'empty', net: 'empty', fs: 'empty' },
       performance: {
@@ -66,7 +66,7 @@ module.exports = (babelEnv) => {
       module: {
         rules: [
           {
-            test: /\.jsx?$/,
+            test: /[jt]sx?$/,
             include: [path.join(packageRoot, 'src'), path.join(packageRoot, 'node_modules')],
             use: [babelLoader(babelEnv)],
           },