+++ packages/one-app-dev-bundler/esbuild/plugins/cjs-compatibility-hotpatch.js
@@ -27,6 +27,9 @@ const cjsCompatibilityHotpatch = {
   name: 'cjsCompatibilityHotpatch',
   setup(build) {
     build.onEnd(async (result) => {
+      if (!result.metafile) {
+        return result;
+      }
       const fileNames = getJsFilenamesFromKeys(result.metafile.outputs);
       await Promise.all(fileNames.map(async (fileName) => {
         const initialContent = await fs.promises.readFile(fileName, 'utf8');