+++ packages/holocron-dev-server/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.14](https://github.com/americanexpress/one-app-cli/compare/@americanexpress/holocron-dev-server@0.1.13...@americanexpress/holocron-dev-server@0.1.14) (2023-06-13)
+
+**Note:** Version bump only for package @americanexpress/holocron-dev-server
+
+
+
+
+
+## [0.1.13](https://github.com/americanexpress/one-app-cli/compare/@americanexpress/holocron-dev-server@0.1.12...@americanexpress/holocron-dev-server@0.1.13) (2023-05-15)
+
+**Note:** Version bump only for package @americanexpress/holocron-dev-server
+
+
+
+
+
+## [0.1.12](https://github.com/americanexpress/one-app-cli/compare/@americanexpress/holocron-dev-server@0.1.10...@americanexpress/holocron-dev-server@0.1.12) (2023-05-03)
+
+**Note:** Version bump only for package @americanexpress/holocron-dev-server
+
+
+
+
+
+## [0.1.11](https://github.com/americanexpress/one-app-cli/compare/@americanexpress/holocron-dev-server@0.1.10...@americanexpress/holocron-dev-server@0.1.11) (2023-04-07)
+
+**Note:** Version bump only for package @americanexpress/holocron-dev-server
+
+
+
+
+
 ## [0.1.10](https://github.com/americanexpress/one-app-cli/compare/@americanexpress/holocron-dev-server@0.1.9...@americanexpress/holocron-dev-server@0.1.10) (2023-03-16)
 
 **Note:** Version bump only for package @americanexpress/holocron-dev-server