+++ packages/one-app-bundler/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [6.20.0](https://github.com/americanexpress/one-app-cli/compare/@americanexpress/one-app-bundler@6.19.1...@americanexpress/one-app-bundler@6.20.0) (2023-06-13)
+
+
+### Features
+
+* **purgecss:** allow for backwards compatibility of purgecss 2 and 3 ([#519](https://github.com/americanexpress/one-app-cli/issues/519)) ([3ab7d58](https://github.com/americanexpress/one-app-cli/commit/3ab7d583957b02093868c104137a90f4a3e27242))
+
+
+
+
+
+## [6.19.1](https://github.com/americanexpress/one-app-cli/compare/@americanexpress/one-app-bundler@6.19.0...@americanexpress/one-app-bundler@6.19.1) (2023-05-15)
+
+**Note:** Version bump only for package @americanexpress/one-app-bundler
+
+
+
+
+
+# [6.19.0](https://github.com/americanexpress/one-app-cli/compare/@americanexpress/one-app-bundler@6.18.1...@americanexpress/one-app-bundler@6.19.0) (2023-05-03)
+
+
+### Bug Fixes
+
+* **cyptoHash:** patch crypto hash to support md4 hash ([#516](https://github.com/americanexpress/one-app-cli/issues/516)) ([b3d6961](https://github.com/americanexpress/one-app-cli/commit/b3d696180d62f2e4905f4d9d54d980022a78f1c1))
+
+
+### Features
+
+* **typeScript:** allow bundlers to consume ts files ([#520](https://github.com/americanexpress/one-app-cli/issues/520)) ([77cb796](https://github.com/americanexpress/one-app-cli/commit/77cb796ed028cae25d3fcc38171d84594d15e3d3))
+
+
+
+
+
+## [6.18.2](https://github.com/americanexpress/one-app-cli/compare/@americanexpress/one-app-bundler@6.18.1...@americanexpress/one-app-bundler@6.18.2) (2023-04-07)
+
+**Note:** Version bump only for package @americanexpress/one-app-bundler
+
+
+
+
+
 ## [6.18.1](https://github.com/americanexpress/one-app-cli/compare/@americanexpress/one-app-bundler@6.18.0...@americanexpress/one-app-bundler@6.18.1) (2023-03-16)
 
 **Note:** Version bump only for package @americanexpress/one-app-bundler